refactor(workflow): split commit.go into 3 focused files
